+++ team-builder/src/constants/roles.js
@@ -1,0 +1,11 @@
+const roles = [
+  "Front End Engineer",
+  "Back End Engineer",
+  "Designer",
+  "Junior Front End Developer",
+  "Omnipotent Unicorn",
+  "Wise Old Dragon",
+  "Aspiring Sith Master",
+];
+
+export default roles;
